% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_boundary_layer.R
\name{add_boundary_layer}
\alias{add_boundary_layer}
\title{Add a boundary layer to a deck}
\usage{
add_boundary_layer(rdeck, id, name, data, ...)
}
\arguments{
\item{rdeck}{an rdeck object}

\item{id}{a character string specifying the id of the layer to add}

\item{name}{a character string specifying the name of the layer to add}

\item{data}{a spatial object containing the boundary data to add}

\item{...}{additional arguments to modify the layer appearance or behavior, passed to \code{add_polygon_layer}.}
}
\value{
a deck object with the new boundary layer added
}
\description{
This function adds a boundary layer to a deck in the form of a R deck object.
The function takes an id, name, and spatial data, and creates a polygon layer that can be added to the deck.
Uses the 'sf' package to transform the spatial data to WGS84 CRS and extract the polygon geometry.
Additional arguments can be passed to modify the layer appearance or behavior.
}
