% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_layer.R
\name{add_data_centroid_layer}
\alias{add_data_centroid_layer}
\title{Add a data centroid layer to the rdeck map}
\usage{
add_data_centroid_layer(rdeck, id, visible = FALSE, dataset_num, ...)
}
\arguments{
\item{rdeck}{an rdeck object}

\item{id}{a character string specifying the id of the layer to be added}

\item{visible}{a logical value indicating whether the layer is visible or not}

\item{dataset_num}{an integer specifying the number of the dataset to be displayed}
}
\value{
the rdeck object with the scatterplot layer added
}
\description{
This function adds the centroid of a dataset to an rdeck map. The function requires
the id of the layer to be added and the number of the dataset to be displayed.
The layer can also be made visible or invisible with the \code{visible} argument.
}
