% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_layer_name.R
\name{check_layer_name}
\alias{check_layer_name}
\title{Check if a layer exists in an rdeck}
\usage{
check_layer_name(rdeck, layer_name)
}
\arguments{
\item{rdeck}{an rdeck object}

\item{layer_name}{a character string specifying the name of the layer to check}
}
\value{
a logical value indicating whether the layer exists in the deck
}
\description{
This function checks if a specified layer name exists in a deck in the form
of an rdeck object.
The function returns a logical value indicating whether the layer name exists
in the deck.
}
