% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_points_to_sfc.R
\name{convert_points_to_sfc}
\alias{convert_points_to_sfc}
\title{Convert a data frame of points to an sf object}
\usage{
convert_points_to_sfc(
  centroid_points = .,
  longitude_col = "longitude",
  latitude_col = "latitude",
  tooltip
)
}
\arguments{
\item{centroid_points}{a data frame containing the points to be converted}

\item{longitude_col}{a character string specifying the name of the longitude column in centroid_points}

\item{latitude_col}{a character string specifying the name of the latitude column in centroid_points}

\item{tooltip}{a character string specifying the tooltip to be added to the converted sf object}
}
\value{
an sf object containing the converted points and the specified tooltip
}
\description{
This function takes a data frame of points and converts it to an sf object.
The function requires the longitude and latitude columns to be specified.
A tooltip can also be added to the converted sf object.
}
