% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_entries.R
\name{extract_entries}
\alias{extract_entries}
\title{Extract entries from a list based on the name}
\usage{
extract_entries(api_list, entry_name, fixed = TRUE)
}
\arguments{
\item{api_list}{a list (probably from the GitHub api)}

\item{entry_name}{the name of the entry to be extracted from the api list e.g. "html_url"}

\item{fixed}{TRUE/FALSE whether or not the exact string should be used.
FALSE would mean regex will be interpretted.}
}
\value{
a subset of api_list that only contains items with the entry_name string in the name.
}
\description{
Given a list, extract items based on their name
}
\examples{

repo_info <- get_repo_info("jhudsl/Documentation_and_Usability")

extract_entries(repo_info, "release", fixed = TRUE)
}
