% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weighted_centroid.R
\name{find_weighted_centroid}
\alias{find_weighted_centroid}
\title{Find the weighted centroid of a given shapefile with provided weights}
\usage{
find_weighted_centroid(shape, longitude, latitude, person_weight)
}
\arguments{
\item{shape}{Object of class \code{sf} that has the coordinates and weights as features.}

\item{longitude}{Column containing the longitude of the centroids to find the
centre of.}

\item{latitude}{Column containing the latitude of the centroids to find the
centre of.}

\item{person_weight}{Column containing the relative weights for each centroid.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{centre_longitude}: Longitude of the geographic centre
\item \code{centre_latitude}: Latitude of the geographic centre
}
}
\description{
Finds the weighted centroid of a given set of coordinates, weighted by a
user-specified column. The centre is calculated as:
\deqn{(x_c, y_c) = \left( \sum_i x_i w_i, \sum_i y_i w_i \right) / \sum_j w_j}
}
\examples{
find_weighted_centroid(wphu_mb_shp, x, y, Person)
}
