% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chapters.R
\name{get_chapters}
\alias{get_chapters}
\title{Retrieve bookdown chapters for a repository}
\usage{
get_chapters(repo_name, git_pat = NULL)
}
\arguments{
\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().
Authorization handled by \link{@get_git_auth}}

\item{org_name}{the name of the organization that e.g. "jhudsl"}

\item{url_base}{the base url of the github pages for this organization
e.g. "http://jhudatascience.org/"}

\item{output_file}{a file path for where the chapter information should be
saved e.g. "jhudsl_chapter.tsv"}
}
\value{
A TRUE/FALSE whether or not the repository exists. Optionally the
output from git ls-remote if return_repo = TRUE.
}
\description{
Given an repository on GitHub, retrieve all the bookdown chapter
information for the Github page if it exists.
Currently only public repositories are supported.
}
\examples{

get_chapters("jhudsl/DaSL_Course_Template_Bookdown")

}
