% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_obj.R
\name{get_learning_obj}
\alias{get_learning_obj}
\title{Retrieve learning objectives for a chapter}
\usage{
get_learning_obj(url, prompt = "This chapter will demonstrate how to\\\\:")
}
\arguments{
\item{url}{a url to a bookdown chapter
e.g. "https://jhudatascience.org/Documentation_and_Usability/what-does-good-documentation-look-like.html"}

\item{prompt}{A string character that is a prompt lead in to the Learning objectives and should be deleted.}
}
\value{
find learning objectives for a particular chapter
}
\description{
Given an repository on GitHub, retrieve all the bookdown chapter
information for the Github page if it exists.
Currently only public repositories are supported.
}
\examples{

# Declare chapter URL
url <- "https://jhudatascience.org/Documentation_and_Usability/other-helpful-features.html"

get_learning_obj(url)
}
