% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_info.R
\name{get_release_info}
\alias{get_release_info}
\title{Retrieve information about a github repo}
\usage{
get_release_info(repo_name, git_pat = NULL, verbose = TRUE)
}
\arguments{
\item{repo_name}{The full name of the repo to get bookdown chapters from.
e.g. "jhudsl/OTTR_Template"}

\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().
Authorization handled by \link[githubr]{get_git_auth}}

\item{verbose}{TRUE/FALSE do you want more progress messages?}
}
\value{
a data frame with the repository's release information: tag_name and tag_date.
NAs are returned in these columns if there are no releases.
}
\description{
Given an repository on GitHub, retrieve the information about it from the
GitHub API and read it into R.
}
\examples{
\dontrun{

release_info <- get_release_info("jhudsl/OTTR_Template")
}
}
