% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_org_chapters.R
\name{retrieve_org_chapters}
\alias{retrieve_org_chapters}
\title{Retrieve all bookdown chapters for an org}
\usage{
retrieve_org_chapters(
  org_name = NULL,
  output_file = "org_chapter_info.tsv",
  git_pat = NULL,
  retrieve_learning_obj = FALSE,
  retrieve_keywords = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{org_name}{the name of the organization that e.g. "jhudsl"}

\item{output_file}{a file path for where the chapter information should be
saved e.g. "jhudsl_chapter.tsv"}

\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().}

\item{retrieve_learning_obj}{TRUE/FALSE attempt to retrieve learning objectives?}

\item{retrieve_keywords}{TRUE/FALSE attempt to retrieve keywords from the chapter?}

\item{verbose}{TRUE/FALSE do you want more progress messages?}
}
\value{
A TRUE/FALSE whether or not the repository exists. Optionally the
output from git ls-remote if return_repo = TRUE.
}
\description{
Given an organization on GitHub, retrieve all the bookdown chapter
information for all the Github pages that exist for all the repositories.
Currently only public repositories are supported.
}
\examples{

retrieve_org_chapters(
  org_name = "jhudsl",
  output_file = "jhudsl_chapter_info.tsv"
)
}
