% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_to_wphu.R
\name{trim_to_wphu}
\alias{trim_to_wphu}
\title{Trim points to within WPHU (Western Public Health Unit)}
\usage{
trim_to_wphu(pts)
}
\arguments{
\item{pts}{a data frame of points containing columns 'longitude' and 'latitude'}
}
\value{
a data frame of points that are within WPHU
}
\description{
This function takes a set of points and determines which points fall within
WPHU (Western Public Health Unit) based on their longitude and latitude
coordinates. The function uses the 'sf' package to create points and check
if they intersect with the outline of the WPHU shapefile provided to the
function. If a point is within the WPHU, it is kept in the final output.
If a point is outside of the WPHU, it is discarded from the final output.
}
\examples{

data(example_lga_person_weights)

trim_to_wphu(example_lga_person_weights)

}
