% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chapters.R
\name{get_chapters}
\alias{get_chapters}
\title{Retrieve bookdown chapters for a repository}
\usage{
get_chapters(
  repo_name,
  git_pat = NULL,
  retrieve_learning_obj = FALSE,
  retrieve_keywords = TRUE,
  verbose = TRUE,
  udmodel = NULL
)
}
\arguments{
\item{repo_name}{The full name of the repo to get bookdown chapters from.
e.g. "jhudsl/DaSL_Course_Template_Bookdown"}

\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().
Authorization handled by \link[cow]{get_git_auth}}

\item{retrieve_learning_obj}{TRUE/FALSE attempt to retrieve learning objectives?}

\item{retrieve_keywords}{TRUE/FALSE attempt to retrieve keywords from the chapter?}

\item{verbose}{TRUE/FALSE do you want more progress messages?}

\item{udmodel}{A udmodel passed in for keyword determination. Will be obtained using
\code{udpipe::udpipe_download_model(language = "english")} if its not given.}
}
\value{
a data frame with the repository with the following columns:
data_level, data_path, chapt_name, url, repository name
}
\description{
Given an repository on GitHub, retrieve all the bookdown chapter
information for the Github page if it exists.
Currently only public repositories are supported.
}
\examples{
\dontrun{

usethis::create_github_token()

get_chapters("jhudsl/Documentation_and_Usability")

}
}
