% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chapters.R
\name{get_chapters}
\alias{get_chapters}
\title{Retrieve bookdown chapters for a repository}
\usage{
get_chapters(repo_name, git_pat = NULL)
}
\arguments{
\item{repo_name}{The full name of the repo to get bookdown chapters from.
e.g. "jhudsl/DaSL_Course_Template_Bookdown"}

\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().
Authorization handled by \link[gitHelpeR]{get_git_auth}}
}
\value{
A TRUE/FALSE whether or not the repository exists. Optionally the
output from git ls-remote if return_repo = TRUE.
}
\description{
Given an repository on GitHub, retrieve all the bookdown chapter
information for the Github page if it exists.
Currently only public repositories are supported.
}
\examples{

get_chapters("jhudsl/DaSL_Course_Template_Bookdown")
}
