% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_org_chapters.R
\name{retrieve_org_chapters}
\alias{retrieve_org_chapters}
\title{Retrieve all bookdown chapters}
\usage{
retrieve_org_chapters(
  org_name = NULL,
  url_base = NULL,
  output_file = "org_chapter_info.tsv",
  git_pat = NULL
)
}
\arguments{
\item{org_name}{the name of the organization that e.g. "jhudsl"}

\item{url_base}{the base url of the github pages for this organization
e.g. "http://jhudatascience.org/"}

\item{output_file}{a file path for where the chapter information should be
saved e.g. "jhudsl_chapter.tsv"}

\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().}
}
\value{
A TRUE/FALSE whether or not the repository exists. Optionally the
output from git ls-remote if return_repo = TRUE.
}
\description{
Given an organization on GitHub, retrieve all the bookdown chapter
information for all the Github pages that exist for all the repositories.
Currently only public repositories are supported.
}
\examples{

retrieve_org_chapters(
  org_name = "jhudsl",
  output_file = "jhudsl_chapter_info.tsv"
)
}
